Add support for progamatically flipping cards
A nohover attribute was added in order to disable the automatic flippin of cards on hover.
Furthermore a JS API (toggle, showBack, etc) was added a flipped property that allows to change the flip state via data-binding or imperatively.

A flipend event will be fired as soon as the transition ends. Furthermre the flip-card will listen for a show event that is fired from child elements and reveal the corresponding side.

The external stylesheet was moved into the html template as external stylesheets have been deprecated.